Create Excel_Column_Autofit
Open a blank Excel Sheet. Open the Visual Basic Editor and paste the code above. Edit the path of the folder in line 5.
This code open all the excel files in the folder one by one and loops through each sheet of the file while doing a autofit to all the columns and selecting the first cell of each sheet. It then selects the first sheet of the workbook and closes the file in a macrofree format